requirements.txt
asgiref==3.9.1
attrs==25.3.0
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
gunicorn==23.0.0
inflection==0.5.1
jsonschema==3.2.0
packaging==25.0
pillow==11.3.0
PyJWT==2.10.1
pyrsistent==0.20.0
PyYAML==6.0.2
setuptools==80.9.0
six==1.17.0
spectacular==0.4.0
sqlparse==0.5.3
uritemplate==4.2.0
